Swap Fixnum for Integer (2.4.0)
As of Ruby 2.4.0, Fixnum and Bignum are both Integer, now. This poses
a bit of a problem in that it's now harder to distinguish whether
a number is a valid fixnum, which was useful here. That said, it'll
just have to be the case that we check the bit length of numbers or
something, since there's nothing in CRuby that helps with this.